Refactor the overall test run code
- move a lot of code out of `specification` into `run.rs`
- move the prerequisite building and git cloning out of test into
  `run.rs`
