+++ src/lib.rs
@@ -9,12 +9,14 @@ pub mod cli_contributor;
 pub mod config;
 pub mod coordinator;
 pub mod drop_participant;
+pub mod frontend;
 pub mod git;
 pub mod join;
 pub mod npm;
 pub mod options;
 pub mod process;
 pub mod reporting;
+pub mod run;
 pub mod rust;
 pub mod specification;
 pub mod state_monitor;
